% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{AddGeneExpressions}
\alias{AddGeneExpressions}
\title{Add Additional Gene Expression Data}
\usage{
AddGeneExpressions(object, expr_data, name_match = "GeneName")
}
\arguments{
\item{object}{FISHprobe object.}

\item{expr_data}{Expression matrix with rownames as gene/transcripts names.}

\item{name_match}{Names used to match the expr_data with object@sequences slot data. Default is GeneName.}
}
\value{
Returns a FISHprobe object.
}
\description{
Add gene expression data for the genes/transcripts used in the object.
}
\examples{
\donttest{
AddGeneExpressions(object, expression_data, name_match = "ensembl_transcript_id")
}
}
