% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{AddReadout}
\alias{AddReadout}
\title{Add Readout Probes}
\usage{
AddReadout(object, readout_data, codebook_data)
}
\arguments{
\item{object}{FISHprobe object.}

\item{readout_data}{Data frame with a set of readout probes.}

\item{codebook_data}{Codebook object with correct format. See \code{\link{FISHprobe-class}} for details.}
}
\value{
Returns a FISHprobe object.
}
\description{
Add readout probes.
}
\examples{
\donttest{
object = AddReadout(object, readout.probes)
}
}
