% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{AssignReadout}
\alias{AssignReadout}
\title{Assign Transcripts/Genes to Readout Probes}
\usage{
AssignReadout(object, channel_transcript, indices = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{channel_transcript}{N channel list contains transcript ids in each channel.}

\item{indices}{N channel list contains specific indices for each transcript in each channel. 
Default is NULL, the transcripts in each channel will be ordered the same as in channel_transcript list.}
}
\value{
Returns a FISHprobe object with updated codebook$Assignment slot.
}
\description{
Assign transcripts/genes to the readout probes/flanks generated.
}
\examples{
\donttest{
channel_list = list("488" = Transcript1, "561" = Transcript2, "647" = Transcript3)
object = AssignReadout(object, channel_transcript = channel_list)
}
}
