% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{BLASTProbes}
\alias{BLASTProbes}
\title{Run BLAST on Probes}
\usage{
BLASTProbes(
  object = NULL,
  sequences = NULL,
  gene_name = NULL,
  transcript_ids = NULL,
  slot = "Target",
  blast_db = NULL,
  repeat_mask = FALSE,
  BLAST_args = "-task blastn-short -evalue 10",
  future_parallel = FALSE,
  future_plan = "multiprocess",
  return_blast = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{sequences}{Sequences to BLAST. Note that return_blast will be set to TRUE for this mode.}

\item{gene_name}{Gene or transcript ID of the input sequences. Default is NULL.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{blast_db}{BLAST database object. See \code{\link{CreateBLASTObject}} for detail. Default is using precomputed human or mouse BLAST databases.
\itemize{
  \item pre-computed human BLAST database object using \href{https://www.gencodegenes.org/human/}{GENCODE} v32 human transcript sequences.
  \item pre-computed mouse BLAST database object using \href{https://www.gencodegenes.org/mouse/}{GENCODE} M23 mouse transcript sequences.
}}

\item{repeat_mask}{Whether to use pre-computed specie specific mask. Default is FALSE. See \code{\link{CreateBLASTObject}} for detail.}

\item{BLAST_args}{Arguments passed to BLAST in command-line style. Default is \code{-task blastn-short -evalue 1},
BLAST for sequences shorter than 50 bases and E value cut-off of 10. See \code{\link[rBLAST]{blast}} for detail.}

\item{future_parallel}{Parallelization with \code{\link[future]{future}} R package. Default is FALSE.}

\item{future_plan}{Parallelization \code{\link[future]{plan}} used. Default is multiprocess.}

\item{return_blast}{Whether to return the BLAST result instead. Default is FALSE.}
}
\value{
Returns a FISHprobe object or BLAST results depends on return_blast choice.
}
\description{
Run BLAST on probe sets against a database. Please install NCBI \href{https://blast.ncbi.nlm.nih.gov/Blast.cgi?PAGE_TYPE=BlastDocs&DOC_TYPE=Download}{BLAST} first.
Note the default blastn is performed on both strands, for short sequences (<50nt), with E value cut-off of 1. 
Parallelization is implemented by default using \code{\link[pbmcapply]{pbmclapply}} with process bars.
However when future_parallel is set to true, \code{\link[future]{future}} will be used to spead up, and the process bar will not be shown.
}
\examples{
\donttest{
object = BLASTProbes(object)

# use customized BLAST database 
bl.used = CreateBLASTObject("customized.fa")
object = BLASTProbes(object, blast_db = bl.used)

# return BLAST results instead of FISHprobe object
bl.results = BLASTProbes(object, blast_db = bl.used, return_blast = TRUE)

# BLAST on user defined sequences
bl.results = BLASTProbes(sequences = sequences.use, gene_name = "gene", blast_db = bl.used)
}
}
