% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{Calc2ndStruct}
\alias{Calc2ndStruct}
\title{Calculate Probe Secondary Structures}
\usage{
Calc2ndStruct(
  object,
  slot = "Target",
  transcript_ids = NULL,
  method = "viennaRNA",
  pseudoknots = FALSE,
  Na_M = 0.33,
  temperature = 37,
  formamide_percent = 50,
  formamide_factor = 0.62,
  nupack_params = ""
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{method}{Method used to calculate secondary structures, either viennaRNA or nupack. Default is viennaRNA.}

\item{pseudoknots}{Whether to consider pseudoknots. Default is FALSE.}

\item{Na_M}{Na concentration in unit molar [M]. Default is 0.33 (330mM, equal to 2XSSC).}

\item{temperature}{Temperature in degree celsius used. The default is 37 (with adjustment of formamide parameters below).}

\item{formamide_percent}{Formamide concentration in percentage. Default is 50.}

\item{formamide_factor}{Formamide factor for temperature calculation, it will adjust for an increase of temperature celsius in [formamide_factor] * [formamide_percent]. Default is 0.62.}

\item{nupack_params}{Additional parameters for NUPACK. Default is none.}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate secondary structures of the probes. This function utilizes run_RNAfold function from LncFinder package and complexes from NUPACK. 
See help page of \code{\link[LncFinder]{run_RNAfold}} and \href{http://www.nupack.org/downloads/documentation}{NUPACK} for details.
Please make sure you have installed \href{https://www.tbi.univie.ac.at/RNA/#download}{ViennaRNA} or \href{http://nupack.org/downloads/}{NUPACK}.
}
\examples{
\donttest{
object = Calc2ndStruct(object)
}
}
