% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcConsecutive}
\alias{CalcConsecutive}
\title{Calculate Consecutive Repeats}
\usage{
CalcConsecutive(
  object,
  slot = "Target",
  transcript_ids = NULL,
  consecutive_repeats = 5,
  colname_use = "Consecutive"
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{consecutive_repeats}{Find consecutive repeats of N same nucleotides. Default is 5 nt. See \code{\link{ExtractProbes}} for details.}

\item{colname_use}{Column name used to store consecutive repeat counts. Default is Consecutive.}
}
\value{
Returns a FISHprobe object.
}
\description{
Find consecutive repeats of N same nucleotides.
}
\examples{
\donttest{
object = CalcConsecutive(object)
}
}
