% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcDuplex}
\alias{CalcDuplex}
\title{Calculate Duplexing Free Energy and Percentage (Probability)}
\usage{
CalcDuplex(
  object,
  slot = "Target",
  transcript_ids = NULL,
  Na_M = 0.33,
  probe_M = 1e-06,
  target_M = 1e-06,
  temperature = 37,
  formamide_percent = 50,
  formamide_factor = 0.62,
  nupack_params = ""
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{Na_M}{Na concentration in unit molar [M]. Default is 0.33 (330mM, equal to 2XSSC).}

\item{probe_M}{Concentration of the probes in unit molar [M]. Default is 1e-6 (1uM).}

\item{target_M}{Concentration of the targeting sequences in unit molar [M]. Default is 1e-6 (1uM).}

\item{temperature}{Temperature in degree celsius used. The default is 37 (with adjustment of formamide parameters below).}

\item{formamide_percent}{Formamide concentration in percentage. Default is 50.}

\item{formamide_factor}{Formamide factor for temperature calculation, it will adjust for an increase of temperature celsius in [formamide_factor] * [formamide_percent]. Default is 0.62.}

\item{nupack_params}{Additional parameters for NUPACK. Default is none.}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate duplex binding free energy and duplex precetage (can be treated as duplex probability) between probes and targets.
}
\details{
See \href{http://www.nupack.org/downloads/serve_public_file/nupack_user_guide_3.2.2.pdf?type=pdf}{NUPACK 3.2.2 User Guide} for details.
Also see \href{https://www.pnas.org/content/115/10/E2183}{OligoMiner} for details.
}
\examples{
\donttest{
object = CalcDuplex(object, formamide_percent = 20)
}
}
