% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcExon}
\alias{CalcExon}
\title{Calculate Exon Information}
\usage{
CalcExon(object, transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate exon information of the probes.
Please note the numbering of the exons may be different from transcript to transcript of the same gene. 
For example Exon-5 can have different sequences between two isoforms.
}
\examples{
\donttest{
object = CalcJunction(object)
}
}
