% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcGC}
\alias{CalcGC}
\title{Calculate GC Content}
\usage{
CalcGC(object, slot = "Target", transcript_ids = NULL, colname_use = "GC")
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{colname_use}{Column name used to store GC. Default is GC.}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate GC content of probes.
}
\examples{
\donttest{
object = CalcGC(object)
}
}
