% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcRepMask}
\alias{CalcRepMask}
\title{Calculate Repetitive Masks}
\usage{
CalcRepMask(object, transcript_ids = NULL, masks_names = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{masks_names}{Names of built-in masks. Default is NULL (the default active build-in masks, see below).
\itemize{
  \item AGAPS, assembly gaps (big N-blocks), active by default.
  \item AMB, intra-contig ambiguities (IUPAC ambiguity letter), active by default.
  \item RM, RepeatMasker, not active by default. 
  \item TRF, Tandem Repeats Finder, not active by default. 
  \item All, will activate all above masks.
  }}
}
\value{
Returns a FISHprobe object.
}
\description{
Find the length of overlapping betweem probes and repetitive masked region in the genome.
This length varies between 0 (non-overlapping) to the length of the probe (full-overlapping).
}
\examples{
\donttest{
object = CalcRepMask(object)
}
}
