% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{CalcTm}
\alias{CalcTm}
\title{Calculate Melting Temperatures}
\usage{
CalcTm(
  object,
  slot = "Target",
  transcript_ids = NULL,
  Na = 330,
  probe_nM = 5,
  target_nM = 5,
  formamide_percent = 50,
  formamide_factor = 0.65,
  colname_use = "Tm",
  ...
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{Na}{Na concentration in [mM]. Default is 330mM (equal to 2XSSC).}

\item{probe_nM}{Concentration of the probes in [nM]. Default is 5nM.}

\item{target_nM}{Concentration of the targeting sequences in [nM].. Default is 5nM.}

\item{formamide_percent}{Formamide concentration in percentage. Default is 50.}

\item{formamide_factor}{Coeffecient of Tm decrease per percent formamide.. Default is 0.65.}

\item{colname_use}{Column name used to store Tm. Default is Tm.}

\item{...}{Other arguments can be passed for Tm calculation. See \code{\link[TmCalculator]{Tm_NN}} for details.}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate Tm melting temperatures of probes. This function relies on Tm_NN and chem_correction functions from TmCalculator package.
See \code{\link[TmCalculator]{Tm_NN}} and \code{\link[TmCalculator]{chem_correction}} help page for details.
Note that if Tm is negative after adjusted for formamide etc. it will be reset to zero.
}
\examples{
\donttest{
object = CalcTm(object, Na = 350, probe_nM = 10, target_nM = 10, formamide_percent = 10)
}
}
