% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook_v1}
\alias{codebook_v1}
\title{Pre-Computed Codebook}
\format{An object of class \code{list} of length 4.}
\usage{
codebook_v1
}
\description{
A codebook object with 480 pre-computed 4-bit code in 32 hybridization rounds in 3 channels(488nm, 561nm, 647nm).
}
\keyword{datasets}
