% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{ConcatReadout}
\alias{ConcatReadout}
\title{Concatenate Readout Probe Sequences}
\usage{
ConcatReadout(object, concat_list, insert = "")
}
\arguments{
\item{object}{FISHprobe object.}

\item{concat_list}{A list of indice of concatenated probes.}

\item{insert}{Inserted based between each probes. Default is none.}
}
\value{
Returns a FISHprobe object.
}
\description{
Concatenate readout probe sequences based on their indice.
}
\examples{
\donttest{
object = ConcatReadout(object, concat_list = list(c(1,2),c(1,3),c(2,3)))
}
}
