% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{CreateBLASTObject}
\alias{CreateBLASTObject}
\title{Create BLAST Object}
\usage{
CreateBLASTObject(file = NULL, specie = NULL, masked_repeats = FALSE)
}
\arguments{
\item{file}{FASTA file path and name.}

\item{specie}{Either \emph{human} or \emph{mouse}. Load pre-computed blast database object.}

\item{masked_repeats}{Whether to load pre-computed specie specific repetitive sequences. Default is FALSE.}
}
\value{
Returns a BLAST database object.
}
\description{
Create BLAST object from a FASTA file or load pre-computed BLAST database. See \code{\link[rBLAST]{makeblastdb}} and \code{\link[rBLAST]{blast}} for detail.
}
\details{
Pre-computed human BLAST database object was built using \href{https://www.gencodegenes.org/human/}{GENCODE} v31 human transcript sequences.
Pre-computed mouse BLAST database object was built using \href{https://www.gencodegenes.org/mouse/}{GENCODE} M22 mouse transcript sequences.
Pre-computed repetitive masked sequence BLAST database object was built from \href{https://genome.ucsc.edu/cgi-bin/hgTables}{UCSC}.
}
\examples{
\donttest{
# generate customized BLAST database 
bl.used = CreateBLASTObject("customized.fa")
}
}
