% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{CreateFASTAFile}
\alias{CreateFASTAFile}
\title{Write to FASTA File}
\usage{
CreateFASTAFile(
  object = NULL,
  sequences = NULL,
  transcript_ids = NULL,
  slot = "Target",
  filepath = "temp.fa",
  seq_name = "Seq"
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{sequences}{Sequences to write into FASTA format.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{filepath}{FASTA file path and name. Default is to write to temp.fa in current directory.}

\item{seq_name}{Sequence names or suffix. Default is adding Seq suffix.}
}
\description{
Write sequences into FASTA format.
}
\examples{
\donttest{
probe.data = GetSlotData(object, slotname = "probes", subname = "Target")
CreateFASTAFile(probe.data[[1]][["Sequence"]], filepath = "probe.fa")
}
}
