% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateProbeObject}
\alias{CreateProbeObject}
\title{Create FISH Probe Object}
\usage{
CreateProbeObject(
  gene_names = "",
  specie_type = "human",
  preload_readout = FALSE,
  ensemble_transcript_id = NULL,
  ensemble_gene_id = NULL,
  customized_name = NULL,
  customized_data = NULL
)
}
\arguments{
\item{gene_names}{Genes of interest. Mouse gene symbols can be found in \code{\link{mousegenes}}. Human gene symbols can be found in \code{\link{humangenes}}.}

\item{specie_type}{Either \emph{human} or \emph{mouse}. Default is \emph{human}.}

\item{preload_readout}{Logical value whether to load pre-computed readout probes and codebook. Default is FALSE.}

\item{ensemble_transcript_id}{Ensemble transcript IDs used. Default is None.}

\item{ensemble_gene_id}{Ensemble gene IDs used. Default is None.}

\item{customized_name}{Customized gene name. Default is None.}

\item{customized_data}{Customized sequences used. Default is None.}
}
\value{
Returns a \code{\link{FISHprobe-class}} object.
}
\description{
Creat FISH probe object.
}
\examples{
\donttest{
object = CreateProbeObject("Egfr", specie_type = "mouse")
object = CreateProbeObject(c("EGFR","ACTB"), specie_type = "human")
object = CreateProbeObject(c("EGFR","ACTB"), specie_type = "human")
}
}
