% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{DiffFeature}
\alias{DiffFeature}
\title{Feature Differences Between Two Probe Groups}
\usage{
DiffFeature(object, transcript_ids = NULL, feature, group_by = "Part")
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{feature}{Feature used to compare.}

\item{group_by}{Group of probes. Default is splitted probe groups ("Part").}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate the absolute numerical value differences of selected feature between two different probe groups.
}
\examples{
\donttest{
object = DiffFeature(object, feature = "Tm")
}
}
