% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{DiffJunction}
\alias{DiffJunction}
\title{Differenctial Exon Junction Probes Between Transcript Isoforms}
\usage{
DiffJunction(object, transcripts_selected, transcripts_to_compare = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcripts_selected}{Transcript IDs selected to get the differential sequences.}

\item{transcripts_to_compare}{Transcript IDs to compare to. Default is all other isoforms than transcript_selected from the same gene.}
}
\value{
Returns a FISHprobe object, with "DiffJunction" added to @probes slot with differential, shared or none labels indicating exon junction types.
}
\description{
Get differential or shared Probes that targeting exon junctions between different transcript isoforms.
Note this won't work if the gene spans multiple locations on different chromosomes.
}
\examples{
\donttest{
# ENST00000216037 and ENST00000344347 are isoforms of human XBP1 genes. 
xbp1 = CreateProbeObject("XBP1")
xbp1 = TissueExpr(xbp1)
xbp1 = GetSequence(xbp1)
xbp1_diff = ExtractProbes(xbp1)
object = CalcJunction(object)
object = DiffJunction(object, transcripts_selected = "ENST00000216037", 
                      transcripts_to_compare = "ENST00000344347")
}
}
