% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{DiffProbe-class}
\alias{DiffProbe-class}
\title{The DiffProbe S4 Class}
\description{
The DiffProbe object is a subclass of FISHprobe to design probes for different isoforms.
It has additional slots as follow:
}
\section{Slots}{

\describe{
\item{\code{DiffProbe}}{Data frame of differential or shared sequences between different transcript isoforms.}

\item{\code{DiffProbe.type}}{Type of either sequences either \strong{differential} or \strong{shared}.}
}}

