% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{DiffProbes}
\alias{DiffProbes}
\title{Differenctial Probes Between Transcript Isoforms}
\usage{
DiffProbes(
  object,
  transcripts_selected,
  transcripts_to_compare = NULL,
  find_shared = FALSE,
  probe_length = 25,
  consecutive_repeats = 5
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcripts_selected}{Transcript IDs selected to get the differential sequences.}

\item{transcripts_to_compare}{Transcript IDs to compare to. Default is all other isoforms than transcript_selected from the same gene.}

\item{find_shared}{Whether to find common sequences between isoforms.}

\item{probe_length}{Probe length. Default is 25 nt. See \code{\link{ExtractProbes}} for details.}

\item{consecutive_repeats}{Find consecutive repeats of N same nucleotides. Default is 5 nt. See \code{\link{ExtractProbes}} for details.}
}
\value{
Returns a \code{\link{DiffProbe-class}} object (subclass of \code{\link{FISHprobe-class}}) with differential or shared probes between isoforms.
}
\description{
Get differential or shared Probes between different transcript isoforms.
Note this won't work if the gene spans multiple locations on different chromosomes.
}
\examples{
\donttest{
# ENST00000216037 and ENST00000344347 are isoforms of human XBP1 genes. 
xbp1 = CreateProbeObject("XBP1")
xbp1 = TissueExpr(xbp1)
xbp1 = GetSequence(xbp1)
xbp1_diff = DiffProbes(xbp1, transcripts_selected = "ENST00000216037", 
                       transcripts_to_compare = "ENST00000344347")
}
}
