% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{ExtractProbes}
\alias{ExtractProbes}
\title{Extract Possible Probes}
\usage{
ExtractProbes(object, probe_length = 25, consecutive_repeats = 5)
}
\arguments{
\item{object}{FISHprobe object.}

\item{probe_length}{Probe length. Default is 25 nt.}

\item{consecutive_repeats}{Find consecutive repeats of N same nucleotides. Default is 5 nt.}
}
\value{
Returns a FISHprobe object with all possible probes and information stored in object@probes$Target.
}
\description{
Extract possible probes of a specific length.
}
\examples{
\donttest{
# Generate all possible probes of 30nt long 
# and find all consecutive repeats with 3 same nucleotides.
object = ExtractProbes(object, probe_length = 30, consecutive_repeats = 3)
}
}
