% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{FilterBarplot}
\alias{FilterBarplot}
\title{Plot Filtering Barplot}
\usage{
FilterBarplot(object, slot = "Target", transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\description{
Plot a binary heatmap of filtering conditions for probe seb lection.
}
\examples{
\donttest{
FilterBarplot(object)
}
}
