% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{FilterBLAST}
\alias{FilterBLAST}
\title{Filter BLAST Results}
\usage{
FilterBLAST(
  object = NULL,
  blast_result = NULL,
  slot = "Target",
  transcript_ids = NULL,
  alignment_length = 11,
  E_value_cutoff = NULL,
  filter_by_transcript = FALSE,
  probe_mode = FALSE,
  names_delim = "-",
  names_field = 1,
  parse = TRUE,
  return_blast = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{blast_result}{BLAST result. Note that return_blast will be set to TRUE for this mode.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{alignment_length}{BLAST alignment length cutoff. BLAST hits with alignment length longer than this value will be consider as off-target hits. Default is 11.}

\item{E_value_cutoff}{BLAST alignment E value cutoff. BLAST hits with E value bigger than this will not be removed. Default is None.}

\item{filter_by_transcript}{Whether to filter based on transcript level or gene level. Default is FALSE.}

\item{probe_mode}{Whether to activate probe mode to BLAST the probes against each other. Default is FALSE.}

\item{names_delim}{Choose which delimiter to use to filter the off-taget gene/transcript hits. Default is ; for example ACCTTGGCGGCCAAT;ENSMUST00000000001-1 where ; separates sequences and ID.}

\item{names_field}{Choose which field for gene/transcript ID when probe_mode is on. Eg. For ACCTTGGCGGCCAAT;ENSMUST00000000001-1. the names_field = 1 will give ACCTTGGCGGCCAAT as gene name to filter the off-target probe hits. Default is 1.}

\item{parse}{Whether to parse the BLAST results into genes/transcripts. Default is TRUE}

\item{return_blast}{Whether to return the BLAST result instead. Default is FALSE.}
}
\value{
Returns a FISHprobe object or BLAST results depends on return_blast choice.
}
\description{
Filter BLAST results to get off-target gene/transcript hits.
}
\examples{
\donttest{
object = FilterBLAST(object, alignment_length = 15, E_value_cutoff = 1)

# use customized BLAST results after running BLASTProbes
blast_res = BLASTProbes(sequences = sequences.use, gene_name = "gene", blast_db = bl.used)
blast_filtered = FilterBLAST(object, blast_results = blast_res, alignment_length = 15)
}
}
