% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{FilterHeatmap}
\alias{FilterHeatmap}
\title{Plot Filtering Binary Heatmap}
\usage{
FilterHeatmap(
  object,
  slot = "Target",
  transcript_ids = NULL,
  color = "red",
  show_probe_names = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{color}{Color used for heatmap. Default is red.}

\item{show_probe_names}{Whether to show probe names in the row of heatmap. Default is FALSE. 
Note if set to TRUE when the object contains many probes, the heatmap can be well-covered by the fonnts of the probes.}
}
\description{
Plot a binary heatmap of filtering conditions for probe seb lection.
}
\examples{
\donttest{
FilterHeatmap(object, cols_use = "blue")
}
}
