% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{FilterPairs}
\alias{FilterPairs}
\title{Filter Probe Pairs}
\usage{
FilterPairs(object, transcript_ids = NULL, filter_name)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{filter_name}{Filter table feature in @filter.taable slot used for filtering.}
}
\value{
Returns a FISHprobe object.
}
\description{
Filter pairs of splitted probes to remove singlet probes after filtering.
To make use of this feature \code{\link{FilterProbes}} has to be run first with \code{do.filter = FALSE}.
}
\examples{
\donttest{
object = FilterProbes(object,filter_names = "GC",max = 65,do.filter = FALSE)
object = FilterPairs(object, feature = "Tm")
}

}
