% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{FilterProbes}
\alias{FilterProbes}
\title{Filter Probes}
\usage{
FilterProbes(
  object,
  slot = "Target",
  transcript_ids = NULL,
  filter_names,
  min = -Inf,
  max = Inf,
  equal_to = NULL,
  reverse_select = FALSE,
  do.filter = FALSE,
  warning = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{filter_names}{Names to filtering the probes on. See \code{\link{ListFeatures}} for features names for filtering.}

\item{min}{Minimum values used for filtering. Default is -Inf (one feature).}

\item{max}{Maximum values used for filtering. Default is Inf (one feature).}

\item{equal_to}{Filter probes with features equal to this value. Default is None (if provided will override min and max filters).}

\item{reverse_select}{Whether to perform reverse selection on equal_to, if set TRUE will do "not equal to" operation. Default is FALSE.}

\item{do.filter}{Wether to perform filtering on the probes. Default is FALSE.}

\item{warning}{Wether to show warning message. Default is FALSE.}
}
\value{
Returns a FISHprobe object.
}
\description{
Filter probes based on given pramameters. 
Be cautious for this step, use it as the last step.
If accidental filtering is done please do object@probe=object@stash.probes if probes are stashed first. Then redo filtering again.
}
\examples{
\donttest{
object = FilterProbes(object, filter_names = "GC", min = 50, max = 60)
}
}
