% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{FISHprobe-class}
\alias{FISHprobe-class}
\title{The FISHprobe S4 Class}
\description{
The FISHprobe object stores information associated with the Target and Readout probes, including probe sequences, BLAST results and off-target expression etc. 
The slot information is listed as follow:
}
\section{Slots}{

\describe{
\item{\code{genes}}{Gene names of interest.}

\item{\code{transcripts}}{List of ENSEMBL transcript ids associated with genes slot.}

\item{\code{sequences}}{Gene sequences associated with genes slot.}

\item{\code{sequence.type}}{Type of gene sequences in sequences slot.}

\item{\code{probes}}{List contains Target and Readout probes.}

\item{\code{specie}}{Specie either \emph{mouse} or \emph{human}.}

\item{\code{expr.data}}{List of expression data associated with genes slot.}

\item{\code{off.target.expr}}{List contains off-target expression data.}

\item{\code{blast.data}}{List contains BLAST results.}

\item{\code{filter.table}}{List contains probe filtering conditions.}

\item{\code{stash.probes}}{List contains stashed probe sets.}

\item{\code{codebook}}{List of dataframes contains the binary codebook, indices, conversion table and readout probe sequences used.
\itemize{
   \item The $Binary codebook is a data frame with shape: N of total encoded genes by 1+M of total hybridization rounds.
   \item The $Indice table is a data frame with shape: N of total encoded genes by 1+K of total readout probe used per FISH probe construct.
   \item The $Conversion table is a data frame with shape: N of total encoded genes by 3 (Channel of used, Index of readout probes, Bit assigned to that channel).
   \item The $Sequence is a data frame with shape: N of total encoded genes by 1+K of total readout probe used per FISH probe construct.
   \item The $Assignment is a data frame with final pairs of readout flanks and their assignment to each of the transcript or gene.
   \item All the above codebook slots must have the first column reserved to fluorescent Channel. 
 }}

\item{\code{final.probes}}{Data frame contains the final readout/target design of FISH probes.}

\item{\code{param.log}}{List of command parameter used.}
}}

