% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{GenerateFinalProbes}
\alias{GenerateFinalProbes}
\title{Generate Final Probe Set}
\usage{
GenerateFinalProbes(object, transcript_ids = NULL, concat = "")
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{concat}{Sequence to be added in between readout and target probes. Default is none.}
}
\value{
Returns a FISHprobe object with updated final.probes slot.
}
\description{
Generate the final probe set.
}
\examples{
\donttest{
object = GenerateFinalProbes(object, concat = "T")
}
}
