% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{GenerateProbes}
\alias{GenerateProbes}
\title{Generate Probe Sequences}
\usage{
GenerateProbes(object, slot = "Target", transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\value{
Returns a FISHprobe object.
}
\description{
Generate reverse complementary probe sequences against the targeted sequences.
}
\examples{
\donttest{
object = GenerateProbes(object)
}
}
