% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{GenerateProbeTracks}
\alias{GenerateProbeTracks}
\title{Generate Genome Browser with Probe Tracks}
\usage{
GenerateProbeTracks(
  object = NULL,
  gene = NULL,
  transcript_ids = NULL,
  seqeuences = NULL,
  specie = "human",
  BLAST_args = "-task blastn-short",
  file_name = NULL,
  sort = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{gene}{Gene symbols. Either Target or Readout. Default is NULL.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{seqeuences}{Probe sequences. Default is NULL.}

\item{specie}{Either human or mouse. Default is human.}

\item{BLAST_args}{Arguments passed to BLAST in command-line style. Default is -task blastn-short. See  \code{\link{BLASTProbes}} for detail.}

\item{file_name}{Additional file name for genome browser folder. Default is none.}

\item{sort}{Sort probes by targeting positions in increasing order.}
}
\value{
Returns a FISHprobe object.
}
\description{
Generate genome browsers with probe tracks. For probe set over 1000 probes please split it with \code{\link{SubsetProbes}} first.
}
\details{
If a FISHprobe object is given, it will generate genome browser files for all the genes and transcripts in the object. 
A folder contains the genome browser files for each gene will be created in the current directory with the name of \emph{GeneName-TranscriptID}.

If no FISHprobe object is given, instead gene, transcript_id and seqeuences are given, it will generate genome browser for the input gene/transcript and probe sequences.
A genome browser folder named as \emph{GeneName} will be created.
Note that under this mode the probe tracks are generated using BLAST and the probes contain exon-exon junction may show multiple bands (>2) instead of 2.

The genome browser is created using D3GB R package, in order to visualize it with your local browser please refer to this \href{ http://d3gb.usal.es/help/index.html}{page}.
}
\examples{
\donttest{
GenerateProbeTracks(object)
# generate customized genome browser 
GenerateProbeTracks(gene = "Actb",transcript_id = "ENSMUSG00000029580",
                    specie = "mouse",seqeuences = ProbeSequences)
}
}
