% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{GenerateReadout}
\alias{GenerateReadout}
\title{Generate Readout Probes}
\usage{
GenerateReadout(
  object,
  scheme = list(1:2, 3:4),
  flank_5end = "",
  flank_3end = "",
  concat = ""
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{scheme}{Scheme for concatenate readout probes. A list contains 2 set of readout probes to be combined. Default 1-2 and 3-4 position combination.}

\item{flank_5end}{Sequence to be added to the 5' end flank regions. Default is none.}

\item{flank_3end}{Sequence to be added to the 3' end flank regions. Default is none.}

\item{concat}{Sequence to be added in between readout probes. Default is none.}
}
\value{
Returns a FISHprobe object with generated readout probes/flanks in the codebook$Assignment slot.
}
\description{
Generate readout probes/flanks based on codebook and readout probes.
}
\examples{
\donttest{
object = GenerateReadout(object, flank = "AAAAAAAA", concat = "T")
}
}
