% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GetSequence}
\alias{GetSequence}
\title{Get Transcript Sequences}
\usage{
GetSequence(object, transcript_ids = NULL, seqtype = "cdna")
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{seqtype}{Type of sequences. See \code{\link[biomaRt]{getSequence}} for all possible sequence types. Default is cdna sequences.}
}
\value{
Returns a FISHprobe object.
}
\description{
Get transcript sequences.
I noted that the Ensembl server can be unstable for some reasons, which will result in returning a null gene sequence.
A warning will ring and manual curation may be needed in such cases.
}
\examples{
\donttest{
# Automatic selection of the most expressed transcript
object = GetSequence(object)
# Select by at least 10 logTPM expression in lung tissue
object = SelectTranscript(object, genes = "ACTB", tissue = "lung", TPM_min = 10)
}
}
