% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{GetSlotData}
\alias{GetSlotData}
\title{Get Slot Data}
\usage{
GetSlotData(object, slotname, subname = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slotname}{Slot name.}

\item{subname}{Sub-slot name if there is any.}
}
\value{
Returns the slot data.
}
\description{
Get data from a certain slot.
}
\examples{
\donttest{
# get the target probe data
GetSlotData(object, slotname = "probes", subname = "Target")
}
}
