% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{GridSelect}
\alias{GridSelect}
\title{Grid Search for Optimal Probe Filtering}
\usage{
GridSelect(
  object,
  slot = "Target",
  transcript_ids = NULL,
  features_list,
  target_number = 30,
  do.filter = FALSE,
  order_by_target = TRUE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{features_list}{List of feature names with min and max to filter, similar to \code{\link{FilterProbes}}}

\item{target_number}{Targeted number of probes. Default is 30 probes.}

\item{do.filter}{Wether to perform filtering on the probes. Default is FALSE.}

\item{order_by_target}{Order according to the closeness to target_number. If set FALSE, the filtering will be order by number from high to low. Default is TRUE.}
}
\value{
Returns a data frame of grid search results or a FISHprobe object (depends on do.filter).
}
\description{
Grid search for best parameters to filter the probes.
}
\examples{
\donttest{
# Create a feature list with feature names and min/max values
feature.list = list(BLAST_Hits = list(min = c(5,10)),GC = list(min = c(.4,.5), max = c(.5,.7)))
search.res = GridSelect(object, features_list = feature.list, target_number = 25)
print(search.res)

# Perform filtering on the object
object.new = GridSelect(object, features_list = feature.list, target_number = 25, do.filter = TRUE)
}
}
