% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_mean_human}
\alias{gtex_mean_human}
\title{Human Transcript Mean Tissue Specific Expression (GTEx Analysis V8)}
\format{An object of class \code{matrix} with 199166 rows and 54 columns.}
\usage{
gtex_mean_human
}
\description{
An expression matrix of human ENSEMBL transcript IDs and their mean GTEX tissue specific expressions in [log2TPM].
Rows are transcript IDs and Columns are GTEX tissues.
Check \href{https://gtexportal.org/home/datasets}{GTEX} datasets for more details.
}
\keyword{datasets}
