% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_mouse_meta}
\alias{gtex_mouse_meta}
\title{Mouse ENCODE RNA-seq Metadata}
\format{An object of class \code{data.frame} with 586 rows and 53 columns.}
\usage{
gtex_mouse_meta
}
\description{
A data frame of mouse ENCODE RNA-seq data used for \code{\link{gtex_mean_mouse}}.
Check \href{https://www.encodeproject.org/matrix/?type=Experiment&status=released&replicates.library.biosample.donor.organism.scientific_name=Mus+musculus&assay_slims=Transcription}{ENCODE} datasets for more details.
}
\keyword{datasets}
