% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_pct_human}
\alias{gtex_pct_human}
\title{Human Transcript Tissue Specific Expression Percentage (GTEx Analysis V8)}
\format{An object of class \code{matrix} with 199166 rows and 54 columns.}
\usage{
gtex_pct_human
}
\description{
A matrix of human transcript tissue expression percentage using cutoff of 0 logTPM as expressed.
Rows are transcript IDs and Columns are GTEX tissues.
Check \href{https://gtexportal.org/home/datasets}{GTEX} datasets for more details.
}
\keyword{datasets}
