% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_pct_mouse}
\alias{gtex_pct_mouse}
\title{Mouse Transcript Tissue Specific Expression Percentage}
\format{An object of class \code{matrix} with 103639 rows and 51 columns.}
\usage{
gtex_pct_mouse
}
\description{
A matrix of mouse transcript tissue expression percentage using cutoff of 0 logTPM as expressed.
Rows are transcript IDs and Columns are ENCODE tissues.
Check \href{https://www.encodeproject.org/matrix/?type=Experiment&status=released&replicates.library.biosample.donor.organism.scientific_name=Mus+musculus&assay_slims=Transcription}{ENCODE} datasets for more details.
Check meta data using \code{\link{gtex_mouse_meta}}
}
\keyword{datasets}
