% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human}
\alias{human}
\title{Human BioMart Object GRCh38(hg38)}
\format{An object of class \code{Mart} of length 1.}
\usage{
human
}
\description{
A human biomart object. See \code{\link[biomaRt]{useMart}} for details
}
\keyword{datasets}
