% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_ensemble}
\alias{human_ensemble}
\title{Human Gene Names and ENSEMBLE Gene ID Conversion GRCh38(hg38)}
\format{An object of class \code{data.frame} with 66839 rows and 2 columns.}
\usage{
human_ensemble
}
\description{
A data frame of human gene names and their ENSEMBL gene IDs (GR).
}
\keyword{datasets}
