% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humangenes}
\alias{humangenes}
\title{Human Gene Names}
\format{An object of class \code{data.frame} with 37568 rows and 1 columns.}
\usage{
humangenes
}
\description{
A data frame of human gene names.
}
\keyword{datasets}
