% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{ListFeatures}
\alias{ListFeatures}
\title{List All Possible Features to Filter Probes}
\usage{
ListFeatures(
  object,
  slot = "Target",
  transcript_ids = NULL,
  pattern = "",
  use.stash = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{pattern}{Feature patterns to match, examples include GC, Tm etc. Default is none.}

\item{use.stash}{Whether to use stash.probes data. Default is FALSE.}
}
\value{
Returns a list of possible features to filter on.
}
\description{
List all possible features to filter the probes for each transcript.
}
\examples{
\donttest{
ListFeatures(object)
}
}
