% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{ListTissues}
\alias{ListTissues}
\title{List All Possible Tissues}
\usage{
ListTissues(object, pattern = "")
}
\arguments{
\item{object}{FISHprobe object.}

\item{pattern}{Feature patterns to match, examples include Brain, Colon for human etc. Default is none.}
}
\value{
Returns a list of possible tissues.
}
\description{
List All possible tissues in the object.
}
\examples{
\donttest{
ListTissues(object)
ListTissues(object, pattern = "Brain")
}
}
