% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{MergeProbes}
\alias{MergeProbes}
\title{Merge Probe Objects}
\usage{
MergeProbes(object_list)
}
\arguments{
\item{object_list}{A list of FISHprobe objects.}
}
\value{
Returns a merged FISHprobe object.
}
\description{
Merge a list of FISHprobe objects. Please do not merge object with empty probe set.
}
\examples{
\donttest{
object_merged = MergeProbes(list(object1, object2))
}
}
