% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{Mouse BioMart Object GRCh38(mm10)}
\format{An object of class \code{Mart} of length 1.}
\usage{
mouse
}
\description{
A mouse biomart object. See \code{\link[biomaRt]{useMart}} for details
}
\keyword{datasets}
