% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_ensemble}
\alias{mouse_ensemble}
\title{Mouse Gene Names and ENSEMBLE Gene ID Conversion GRCh38(mm10)}
\format{An object of class \code{data.frame} with 56399 rows and 2 columns.}
\usage{
mouse_ensemble
}
\description{
A data frame of mouse gene names and their ENSEMBL gene IDs.
}
\keyword{datasets}
