% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mousegenes}
\alias{mousegenes}
\title{Mouse Gene Names}
\format{An object of class \code{data.frame} with 54563 rows and 1 columns.}
\usage{
mousegenes
}
\description{
A data frame of mouse gene names.
}
\keyword{datasets}
