% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{OffTargetExpr}
\alias{OffTargetExpr}
\title{Calculate Probe Off-Target Expressions}
\usage{
OffTargetExpr(
  object,
  slot = "Target",
  tissue = "All",
  transcript_ids = NULL,
  transcripts_to_exclude = NULL,
  genes_to_exclude = NULL
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{tissue}{Tissue of interest. Default is the mean of all available tissues. Check available tissues with \code{\link{ListTissues}}.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{transcripts_to_exclude}{Transcript IDs to exclude in the off-target BLAST filtering. Default is NULL.}

\item{genes_to_exclude}{Gene names to exclude in the off-target BLAST filtering. Default is NULL.}
}
\value{
Returns a FISHprobe object.
}
\description{
Calculate off-target expression for each probe.
}
\details{
If both gene names and transcript IDs are provided to exclude, the gene names will override.
}
\examples{
\donttest{
object = OffTargetExpr(object)
}
}
