% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{OrderPairs}
\alias{OrderPairs}
\title{Order Probe Pairs}
\usage{
OrderPairs(
  object,
  transcript_ids = NULL,
  feature,
  feature_target = NULL,
  decrease = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{feature}{Feature name used for ordering also see \code{\link{ListFeatures}}.}

\item{feature_target}{Targeting numerical values to order the probes before selection (order probes by the distance to this value). Default is none.}

\item{decrease}{Whether to order probes in decreasing order. Default is FALSE.}
}
\value{
Returns a FISHprobe object.
}
\description{
Order pairs of splitted probes based on feature values.
}
\examples{
\donttest{
object = OrderPairs(object, feature = "Tm")
}

}
