% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{OrderProbe}
\alias{OrderProbe}
\title{Order Probes}
\usage{
OrderProbe(
  object,
  slot = "Target",
  transcript_ids = NULL,
  order_by = "Start",
  decrease = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{order_by}{Parameter to order the probes. Any columns names in object@probes and object@off.target.expr can be used such as GC, Start etc. Default is -5.}

\item{decrease}{Whether order by decreasing order. Default is FALSE.}
}
\value{
Returns a FISHprobe object.
}
\description{
Order probes by a given parameter (starting/end positions, GC content etc).
}
\examples{
\donttest{
# order probes by their starting positions in increasing order
object = OrderProbe(object, order_by = "Start")
# order probes by their GC contents in decreasing order
object = OrderProbe(object, order_by = "GC", decrease = TRUE)
}
}
