% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{PairSelect}
\alias{PairSelect}
\title{Select Pair Probes}
\usage{
PairSelect(
  object,
  transcript_ids = NULL,
  TPM_cutoff = 15,
  n_space = 0,
  order_by_feature = NULL,
  feature_target = NULL,
  iterative = TRUE,
  recalculate = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{TPM_cutoff}{Off-target expression cutoff in [logTPM]. Default is 15.}

\item{n_space}{Spaces between probes. Negative value results in overlapping probes. Default is 0 (non-overlapping).}

\item{order_by_feature}{Whether to select probes based on a numerical feature in \code{\link{ListFeatures}}. Default is none.
Examples including GC content ordering (order probes by the distance to feature_target value).}

\item{feature_target}{Targeting numerical values to order the probes before selection (order probes by the distance to this value). Default is none.}

\item{iterative}{Whether to use iterative mode. Default is TRUE.}

\item{recalculate}{Whether to reselect the probes from stash.probes. Default is FALSE.}
}
\value{
Returns a FISHprobe object.
}
\description{
Select probe prairs after \code{\link{SplitProbes}}, similar as the implementation of \code{\link{SelectProbes}}.
}
\examples{
\donttest{
object = PairSelect(object, n_space = 10)
}

}
