% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotFeatures}
\alias{PlotFeatures}
\title{Barplot for Selected Features}
\usage{
PlotFeatures(
  object,
  features,
  slot = "Target",
  transcript_ids = NULL,
  plot_type = "dot"
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{features}{Features to plot on. See \code{\link{ListFeatures}} for features names for filtering.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{plot_type}{Plot type: "dot" for dot plot, "bar" for bar plot, "hist" for histogram. Default is dot plot.}
}
\description{
Generate barplots for selected features
}
\examples{
\donttest{
PlotFeatures(object,"GC", plot_type = "bar")
}
}
