% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{PlotTissueExpr}
\alias{PlotTissueExpr}
\title{Plot Tissue Expression of Targeting Genes}
\usage{
PlotTissueExpr(object, gene_plot = NULL, return_plot = FALSE)
}
\arguments{
\item{object}{FISHprobe object.}

\item{gene_plot}{Genes of interest to plot. Default is all the genes in the object.}

\item{return_plot}{Return ggplot object. Default is FALSE.}
}
\description{
Plot tissue expressions data after calling \emph{TissueExpr}.
}
\examples{
\donttest{
PlotTissueExpr(object)
PlotTissueExpr(object, gene_plot = "ACTB")
}
}
