% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PrintFilter}
\alias{PrintFilter}
\title{Print Number of Probes After Filtering}
\usage{
PrintFilter(object, slot = "Target", transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\value{
Returns numbers of probes survived after each filtering condition.
}
\description{
Print number of survival probes after different filtering conditions.
}
\examples{
\donttest{
PrintFilter(object)
}
}
