% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readout_v1}
\alias{readout_v1}
\title{Pre-Computed Readout Probe}
\format{An object of class \code{data.frame} with 225 rows and 13 columns.}
\usage{
readout_v1
}
\description{
A set of 226 pre-computed 20mer readout probes
}
\keyword{datasets}
