% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{ResetProbes}
\alias{ResetProbes}
\title{Reset Probe Data}
\usage{
ResetProbes(object, slot = "Target", transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\value{
Returns a FISHprobe object.
}
\description{
Reset probe data to previously stashed version.
}
\examples{
\donttest{
ResetProbes(object)
}
}
