% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{SaveProbes}
\alias{SaveProbes}
\title{Save Probe Sets}
\usage{
SaveProbes(object, slot = "Target", transcript_ids = NULL, suffix = "")
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{suffix}{Additional suffix to be added to the probe file name. Default is none.}
}
\description{
Save probe sets as .csv files with file names as <TranscriptID>_probes.csv.
}
\examples{
\donttest{
SaveProbes(objcet)
}
}
