% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{SelectTranscript}
\alias{SelectTranscript}
\title{Select Transcripts}
\usage{
SelectTranscript(
  object,
  genes = NULL,
  tissue = "All",
  select_best = TRUE,
  TPM_min = -Inf,
  TPM_max = Inf,
  percent_min = -Inf,
  percent_max = Inf
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{genes}{Genes of interest. Default is all the genes in the genes slot.}

\item{tissue}{Tissue of interest. Default is the mean of all available tissues. Check available tissues with \code{\link{ListTissues}}.}

\item{select_best}{Select the most expressed transcript id. Default is TRUE.}

\item{TPM_min}{Minimum logTPM expression level.}

\item{TPM_max}{Maximum logTPM expression level.}

\item{percent_min}{Minimum percentage of expression among tissue replicates.}

\item{percent_max}{Maximum percentage of expression among tissue replicates.}
}
\value{
Returns a FISHprobe object with the selected transcripts for each gene stored in object@transcripts$Selected.
}
\description{
Select transcripts of genes of interest by tissue specific expressions.
}
\examples{
\donttest{
# Automatic selection of the most expressed transcript
object = SelectTranscript(object)
# Select by at least 10 logTPM expression in lung tissue
object = SelectTranscript(object, genes = "ACTB", tissue = "lung", TPM_min = 10)
}
}
