% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{SetSlotData}
\alias{SetSlotData}
\title{Set Slot Data}
\usage{
SetSlotData(object, data, slotname, subname = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{data}{Data to set with.}

\item{slotname}{Slot name.}

\item{subname}{Sub-slot name if there is any.}
}
\value{
Returns a FISHprobe object.
}
\description{
Set data for a certain slot.
}
\examples{
\donttest{
object = SetSlotData(object, data = data.use, slotname = "probes", subname = "Target")
}
}
