% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{SplitProbes}
\alias{SplitProbes}
\title{Split Probes into Two Sub Sequences}
\usage{
SplitProbes(object, transcript_ids = NULL, min_length = 10, nick_length = 0)
}
\arguments{
\item{object}{FISHprobe object.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{min_length}{Minimal sub-probe length in nucleotides. Default is 10 nt.}

\item{nick_length}{Nick region length in nucleotides. Default is 0 nt.}
}
\value{
Returns a new FISHprobe object with splitted probes.
}
\description{
Split the current transcript specific targeting probes into two small sub-sequences with anticipated lengths.
New probes will have a new feature named "Part" either A or B indicating the first and second sub-sequences. 
Note that this will add to the previously calculated variable names with an "Original_" prefix such as "Original_GC".
Therefore you have to recalculate GC content, Tm etc. for each of the newly splitted probes.
}
\examples{
\donttest{
object = SplitProbes(object, min_length = 15, nick_length = 1)
}
}
