% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{StashProbes}
\alias{StashProbes}
\title{Stash Probe Sets}
\usage{
StashProbes(object, slot = "Target", transcript_ids = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}
}
\description{
Stash probe sets into stash.probes slot.
}
\examples{
\donttest{
StashProbes(object)
}
}
