% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{SubsetProbes}
\alias{SubsetProbes}
\title{Subset Probe by Sequence}
\usage{
SubsetProbes(
  object,
  slot = "Target",
  transcript_ids = NULL,
  gene_names = NULL,
  sequences = NULL,
  indices = NULL,
  reverse_select = FALSE
)
}
\arguments{
\item{object}{FISHprobe object.}

\item{slot}{Probe slot. Either Target or Readout. Default is Target probe set.}

\item{transcript_ids}{Transcript IDs. Default is NULL (using all the available transcripts in the object).}

\item{gene_names}{Gene names to subset. Default is NULL.}

\item{sequences}{Probe targeting sequences (not the probe sequences) to subset.}

\item{indices}{Probe indices to subset.}

\item{reverse_select}{Whether to perform reverse/removal selection. Default is FALSE.}
}
\value{
Returns a FISHprobe object.
}
\description{
Subset FISHprobe object with sequences
}
\details{
The order of subseting is defined as transcript_ids > gene_anmes > sequences, if all of them are provided.
}
\examples{
\donttest{
object = SubsetProbes(object, transcript_ids = "ENST00000216037")
}
}
