% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{TissueExpr}
\alias{TissueExpr}
\title{Get Tissue Expression of Targeting Genes}
\usage{
TissueExpr(object, genes = NULL)
}
\arguments{
\item{object}{FISHprobe object.}

\item{genes}{Which gene to use. Default is all the genes in the object.}
}
\value{
Returns a FISHprobe object.
}
\description{
This function allows you to get tissue specific expression of your genes of interests.
}
\details{
For Mouse genes, the expression profile is ENCODE RNA-seq \code{\link{gtex_mean_mouse}}.
For Human genes, the expression profile is GTEX RNA-seq \code{\link{gtex_mean_human}}.
}
\examples{
\donttest{
TissueExpr(object)
}
}
