% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{Transcripts}
\alias{Transcripts}
\title{Get Transcript IDs}
\usage{
Transcripts(object)
}
\arguments{
\item{object}{FISHprobe object.}
}
\value{
Returns transcript IDs in the object.
}
\description{
Get transcript IDs.
}
\examples{
\donttest{
Transcripts(object)
}
}
