% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcripts_table_human}
\alias{transcripts_table_human}
\title{Human ENSEMBLE Gene ID and Transcript ID GRCh38(hg38)}
\format{An object of class \code{data.frame} with 199166 rows and 2 columns.}
\usage{
transcripts_table_human
}
\description{
A data frame of human ENSEMBL gene ID and transcript IDs.
}
\keyword{datasets}
