% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcripts_table_mouse}
\alias{transcripts_table_mouse}
\title{Mouse ENSEMBLE Gene ID and Transcript ID GRCh38(mm10)}
\format{An object of class \code{data.frame} with 103639 rows and 2 columns.}
\usage{
transcripts_table_mouse
}
\description{
A data frame of mouse ENSEMBL gene ID and transcript IDs.
}
\keyword{datasets}
